fix: connect PeerToPeer linker to existing components
