+++ src/pages/matches/index.tsx
@@ -1,10 +1,28 @@
 import { NextPage } from "next";
 import Head from "next/head";
 import { MatchesList } from "../../components/matches/matchesList/MatchesList";
-import React from "react";
-import { Typography } from "@mui/material";
+import React, { useEffect } from "react";
+import { Alert, Typography } from "@mui/material";
+import { useRouter } from "next/router";
+import { addAccessToken, addRefreshToken } from "../../api/token";
+import { isLoggedIn } from "../../api/auth";
+import Button from "@mui/material/Button";
+import DynamicLink from "../../components/DynamicLink";
+import BL_CONFIG from "../../utils/bl-config";
 
 const MatchesPage: NextPage = () => {
+  const router = useRouter();
+
+  useEffect(() => {
+    const { refresh_token, access_token } = router.query;
+
+    if (typeof refresh_token === "string" && typeof access_token === "string") {
+      addAccessToken(access_token);
+      addRefreshToken(refresh_token);
+      router.replace("/matches", undefined, { shallow: true });
+    }
+  }, [router]);
+
   return (
     <>
       <Head>
@@ -13,7 +31,20 @@ const MatchesPage: NextPage = () => {
       </Head>
       <div style={{ padding: "1rem" }}>
         <Typography variant="h1">Mine overleveringer</Typography>
-        <MatchesList />
+        {isLoggedIn() ? (
+          <MatchesList />
+        ) : (
+          <>
+            <Alert severity="info">
+              Du må logge inn for å se overleveringene dine
+            </Alert>
+            <DynamicLink href={`${BL_CONFIG.blWeb.basePath}overleveringer`}>
+              <Button variant={"contained"} sx={{ mt: "1rem" }}>
+                Logg inn
+              </Button>
+            </DynamicLink>
+          </>
+        )}
       </div>
     </>
   );
